Move collection formatting to separate file
